+++ ns1/__init__.py
@@ -242,21 +242,6 @@ class NS1:
 
         return ns1.rest.alerts.Alerts(self.config)
 
-    def alerting(self):
-        """
-        Return the alerting namespace for accessing alerting features
-
-        :return: Alerting namespace
-        """
-        from ns1.alerting import UsageAlertsAPI
-
-        # Create or reuse the alerting namespace
-        ns = getattr(self, "_alerting_ns", None)
-        if ns is None:
-            ns = type("AlertingNS", (), {})()
-            ns.usage = UsageAlertsAPI(self)
-            setattr(self, "_alerting_ns", ns)
-        return ns
 
     def billing_usage(self):
         """